Refactor: move js util into JsUtil
